refactor: define vulnerability scanner interfaces
